Convert to new Path api
The only remainder of the old path api usage is for:

 - reading permisions (new path/io doesn't support that)
 - grabbing current_dir, home_dir (which are immediately converted
   to new PathBufs).

I've added 2 helpers for home_dir (Path covertion) and split_paths
(return owned paths). They're not public, but I wouldn't object to them
being public if that made sense.

Note that the permisions checking is now a no-op on non-unix systems
due to a lack of an implimentation (AFAIK, windows perms were a no-op
in std anyhow).